+++ instarest/bots/urls.py
@@ -4,4 +4,6 @@ from .views import BotViewSet
 
 router = routers.SimpleRouter()
 router.register(r'', BotViewSet)
+
+app_name = 'bots'
 urlpatterns = router.urls
