+++ instarest/config/urls.py
@@ -12,4 +12,5 @@ urlpatterns = [
         'auth/',
         include('djoser.urls')),
     path('auth/', include('djoser.urls.authtoken')),
+    path('bots/', include('bots.urls', namespace='bots')),
 ]
